Created advertisements app and refactored code
