+++ recommendations-be/advertisements/serializers.py
@@ -1,0 +1,12 @@
+from rest_framework import serializers
+
+from .models import Advertisement
+
+
+class AdvertisementSerializer(serializers.ModelSerializer):
+    company = serializers.StringRelatedField()
+
+    class Meta:
+        model = Advertisement
+        fields = "__all__"
+        read_only_fields = ["company"]
