+++ recommendations-be/backend/urls.py
@@ -8,5 +8,5 @@ urlpatterns = [
     path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
     path("api/users/", include("recommenders.urls")),
     path("api/shop/", include("shop.urls")),
-    path("api/advertisements/", include("companies.urls")),
+    path("api/advertisements/", include("advertisements.urls")),
 ]
