+++ recommendations-be/companies/admin.py
@@ -3,7 +3,7 @@ from django.contrib import admin
 from users.admin import UsersAdmin
 
 from .forms import CompanyAdminForm
-from .models import Advertisement, CompanyAdmin
+from .models import CompanyAdmin
 
 
 class CompanyAdminAdmin(UsersAdmin):
@@ -24,9 +24,4 @@ class CompanyAdminAdmin(UsersAdmin):
     form = CompanyAdminForm
 
 
-class AdvertisementAdmin(admin.ModelAdmin):
-    list_display = ["company", "title", "type", "reward_for_approval"]
-
-
 admin.site.register(CompanyAdmin, CompanyAdminAdmin)
-admin.site.register(Advertisement, AdvertisementAdmin)
